+++ data/src/main/kotlin/com/acon/acon/data/di/RepositoryModule.kt
@@ -8,7 +8,6 @@ import com.acon.acon.data.repository.SpotRepositoryImpl
 import com.acon.acon.data.repository.TokenRepositoryImpl
 import com.acon.acon.data.repository.UploadRepositoryImpl
 import com.acon.acon.data.repository.UserRepositoryImpl
-import com.acon.acon.data.repository.local.LocalProfileRepositoryImpl
 import com.acon.acon.domain.repository.AreaVerificationRepository
 import com.acon.acon.domain.repository.MapRepository
 import com.acon.acon.domain.repository.OnboardingRepository
@@ -17,7 +16,6 @@ import com.acon.acon.domain.repository.SpotRepository
 import com.acon.acon.domain.repository.TokenRepository
 import com.acon.acon.domain.repository.UploadRepository
 import com.acon.acon.domain.repository.UserRepository
-import com.acon.acon.domain.repository.local.LocalProfileRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -75,10 +73,4 @@ internal abstract class RepositoryModule {
     abstract fun bindsMapRepository(
         impl: MapRepositoryImpl
     ): MapRepository
-
-    @Singleton
-    @Binds
-    abstract fun bindsProfileLocalRepository(
-        impl: LocalProfileRepositoryImpl
-    ) : LocalProfileRepository
 }
