+++ domain/src/main/java/com/acon/acon/domain/repository/ProfileRepository.kt
@@ -2,6 +2,8 @@ package com.acon.acon.domain.repository
 
 import com.acon.acon.domain.model.profile.PreSignedUrl
 import com.acon.acon.domain.model.profile.Profile
+import com.acon.acon.domain.type.UpdateProfileType
+import kotlinx.coroutines.flow.Flow
 
 interface ProfileRepository {
     suspend fun fetchProfile(): Result<Profile>
@@ -12,4 +14,9 @@ interface ProfileRepository {
 
     suspend fun updateProfile(fileName: String, nickname: String, birthday: String?): Result<Unit>
 
+    fun getProfileType(): Flow<UpdateProfileType>
+
+    fun updateProfileType(type: UpdateProfileType)
+
+    suspend fun resetProfileType()
 }
