--- domain/src/main/java/com/acon/acon/domain/repository/local/LocalProfileRepository.kt
@@ -1,10 +1,0 @@
-package com.acon.acon.domain.repository.local
-
-import com.acon.acon.domain.type.UpdateProfileType
-import kotlinx.coroutines.flow.Flow
-
-interface LocalProfileRepository {
-    fun getProfileType(): Flow<UpdateProfileType>
-    fun updateProfileType(type: UpdateProfileType)
-    suspend fun resetProfileType()
-}
