+++ feature/profile/src/main/java/com/acon/acon/feature/profile/composable/screen/profileMod/ProfileModViewModel.kt
@@ -6,7 +6,6 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import com.acon.acon.domain.error.profile.ValidateNicknameError
 import com.acon.acon.domain.repository.ProfileRepository
-import com.acon.acon.domain.repository.local.LocalProfileRepository
 import com.acon.acon.domain.type.UpdateProfileType
 import com.acon.acon.feature.profile.composable.component.TextFieldStatus
 import com.acon.acon.feature.profile.composable.type.FocusType
@@ -27,7 +26,6 @@ import javax.inject.Inject
 @HiltViewModel
 class ProfileModViewModel @Inject constructor(
     private val profileRepository: ProfileRepository,
-    private val localProfileRepository: LocalProfileRepository,
     application: Application
 ) : AndroidViewModel(application), ContainerHost<ProfileModState, ProfileModSideEffect> {
 
@@ -398,11 +396,11 @@ class ProfileModViewModel @Inject constructor(
     private fun updateProfile(fileName: String, nickname: String, birthday: String?) = intent {
         profileRepository.updateProfile(fileName, nickname, birthday)
             .onSuccess {
-                localProfileRepository.updateProfileType(UpdateProfileType.SUCCESS)
+                profileRepository.updateProfileType(UpdateProfileType.SUCCESS)
                 postSideEffect(ProfileModSideEffect.NavigateToProfile)
             }
             .onFailure {
-                localProfileRepository.updateProfileType(UpdateProfileType.FAILURE)
+                profileRepository.updateProfileType(UpdateProfileType.FAILURE)
                 postSideEffect(ProfileModSideEffect.NavigateToProfile)
             }
     }
